apps/api/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
psycopg[binary]==3.2.3
pydantic==2.10.6
python-multipart==0.0.9
requests==2.32.3
rq==1.16.2
redis==5.2.1
python-dotenv==1.0.1
